Split classic battle tests
